More packages (#10)
* authentik
* miniflux
* vikunja
* n8n
* stashapp
* technitium
* phpmyadmin
* pgadmin

---------

+++ Server/Installers/MysqlInstaller.cs
@@ -35,8 +35,26 @@ public class MysqlInstaller(DockerService dockerService, Application application
     Resource IInstaller<Mysql>.Install(Mysql contract, ExecutionPlan plan)
     {
         var network = plan.GetResource<DockerNetwork>(contract.NetworkId);
-
         var container = application.DependsOn.OfType<DockerContainer>().First();
+        dockerService.AttachNetwork(network.Name, container.Name).Wait();
+        
+        if (contract.Admin)
+        {
+            return new MysqlDatabase(
+                User: "root",
+                Password: application.DependsOn.OfType<GeneratedPassword>().First().Value, 
+                Database:"", 
+                Host: container.Name
+                )
+            {
+                DependsOn =
+                [
+                    application,
+                    network
+                ]
+            };
+        }
+
         var name = contract.DatabaseName;
         var password = RandomNumberGenerator.GetString(
             "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ1234567890",
@@ -57,7 +75,6 @@ public class MysqlInstaller(DockerService dockerService, Application application
              """
         );
 
-        dockerService.AttachNetwork(network.Name, container.Name).Wait();
 
         return new MysqlDatabase(name, password, name, container.Name)
         {
@@ -72,12 +89,15 @@ public class MysqlInstaller(DockerService dockerService, Application application
     /// <inheritdoc />
     void IUninstaller<MysqlDatabase>.Uninstall(MysqlDatabase resource)
     {
-        RunSql(
-            $"""
-             DROP DATABASE `{resource.Database}`;
-             DROP USER '{resource.User}';
-             """
-        );
+        if (resource.User != "root")
+        {
+            RunSql(
+                $"""
+                 DROP DATABASE `{resource.Database}`;
+                 DROP USER '{resource.User}';
+                 """
+            );
+        }
 
         var network = resource.DependsOn.OfType<DockerNetwork>().FirstOrDefault();
         if (network == null)
